pyproject.toml
[tool.poetry]
name = "pythonproject6"
version = "0.1.0"
packages = [
    { include = "src" }
]
description = "Приложение для анализа банковских операций"
authors = ["valerii volynchikov <avto162@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
pandas = "^2.2.3"
pytest = "8.3.4"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.14.0"
black = "^24.10.0"
isort = "^5.13.2"
openpyxl = "*"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"
pytest = "^8.3.4"


[build-system]
requires = ["poetry-core>=1.9.1"]
build-backend = "poetry.core.masonry.api"


[flake8]
max-line-length = 119
ignore = ["E203", "W503"]
exclude = [".git"]


[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'
